+++ dashboard/app/models/user.rb
@@ -437,9 +437,9 @@ class User < ApplicationRecord
   # Make sure to include these Concerns after we include the default Devise
   # modules, since it's trying to extend some methods added by those modules
   # that would be overridden by them if we included it before.
-  include Devise::Models::CustomLockable
   include Devise::Models::ManualSessionExpiration
   include Devise::DatabaseAuthenticationOverrides
+  include Devise::LockableOverrides
 
   acts_as_paranoid # use deleted_at column instead of deleting rows
